+++ site/cart/index.html
@@ -27,5 +27,5 @@
             </div>
         </div>
     </body>
-    <script src="cart-script.js"></script>
+    <script src="./cart-script.js" type="module"></script>
 </html>
